core-service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
psycopg2-binary==2.9.9